move expand paths back to client, change diffcontext errors
